--- src/components/index.tsx
@@ -1,5 +1,0 @@
-export {default as Article} from './article/Article';
-export {default as Brand} from './brand/Brand';
-export {default as CTA} from './cta/CTA';
-export {default as Feature} from './feature/Feature';
-export {default as Navbar} from './navbar/Navbar';
