+++ src/components/web/WebPageExports.tsx
@@ -1,0 +1,5 @@
+export {default as Article} from '../../pages/article';
+export {default as Brand} from '../../pages/brand';
+export {default as CTA} from '../../pages/cta';
+export {default as Feature} from '../../pages/feature';
+export {default as Navbar} from './WebNavbar';
