--- src/containers/features/Features.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Features = () => {
-  return (
-    <div>
-     Features
-    </div>
-  )
-}
-
-export default Features
