--- src/containers/index.tsx
@@ -1,6 +1,0 @@
-export {default as Blog} from './blog/Blog';
-export {default as Features} from './features/Features';
-export {default as Footer} from './footer/Footer';
-export {default as Header} from './header/Header';
-export {default as Possibility} from './possibility/Possibility';
-export {default as WhatScheduler} from './whatScheduler/WhatScheduler';
