+++ src/pages/index.tsx
@@ -1,24 +1,23 @@
 import React from 'react';
-import { Footer, Blog, Possibility, Features, WhatScheduler, Header } from '../containers';
-import { Brand, CTA, Navbar } from '../components';
-
+import WebNavbar from "@/components/web/WebNavbar";
+import WebHeader from "@/components/web/WebHeader";
+import WebFooter from "@/components/web/WebFooter";
+import { Article, Brand, CTA, Feature } from "@/components/web/WebPageExports";
 
 const Home: React.FC = () => {
-  return (
-    <div className="App">
-      <div className="gradient__bg">
-        <Navbar />
-        <Header />
-      </div>
-      <Brand />
-      <WhatScheduler />
-      <Features />
-      <Possibility />
-      <CTA />
-      <Blog />
-      <Footer />
-    </div>
-  );
+    return (
+        <div className="App">
+            <div className="gradient__bg">
+                <WebNavbar />
+                <WebHeader />
+            </div>
+            <Article />
+            <Brand />
+            <CTA />
+            <Feature />
+            <WebFooter />
+        </div>
+    );
 };
 
 export default Home;
