+++ src/styles/web/navbar.css
@@ -1,6 +1,4 @@
-
-
-  .scheduler_navbar {
+.scheduler_navbar {
     display: flex;
     justify-content: space-between;
     align-items: center;
